requirements.txt
ultralytics
opencv-python
numpy
pandas
streamlit
sqlalchemy
pillow
scikit-learn
easyocr
plotly
pytest
